// workspaces/liferay-one-workspace/client-extensions/liferay-one-custom-element/src/types/listTypeDefinition.ts
/**
 * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

export type ListTypeDefinition = {
	externalReferenceCode: string;
	id: number;
	listTypeEntries: {
		externalReferenceCode: string;
		key: string;
		name: string;
		name_i18n: {
			[key: string]: string;
		};
	}[];
	name: string;
};